+++ COMMO.Network/COMMO.Network.csproj
@@ -19,4 +19,8 @@
     <PackageReference Include="System.Memory" Version="4.4.0-preview1-25305-02" />
   </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\COMMO\COMMO.csproj" />
+  </ItemGroup>
+
 </Project>
