+++ COMMO/HashHelper.cs
@@ -1,4 +1,4 @@
-namespace COMMO.GameServer {
+namespace COMMO {
 
     public static class HashHelper {
         public const int Start = 1610612741;
