+++ COMMO/ReadOnlyArray.cs
@@ -1,10 +1,8 @@
 using System;
 using System.Collections;
 using System.Collections.Generic;
-using System.Linq;
-using Newtonsoft.Json;
 
-namespace COMMO.GameServer {
+namespace COMMO {
 
 	public sealed class ReadOnlyArray<T> : IReadOnlyList<T> {
 		private readonly T[] _items;
@@ -16,32 +14,19 @@ namespace COMMO.GameServer {
 			_items = items;
 		}
 
-		[JsonConstructor]
-		private ReadOnlyArray(IEnumerable<T> items) {
-			if (items == null)
-				throw new ArgumentNullException(nameof(items));
-
-			_items = items.ToArray();
-		}
-
 		public T this[int index] => _items[index];
 
 		public int Count => _items.Length;
 
-		public IEnumerator<T> GetEnumerator() {
-			return _items.GetGenericEnumerator();
-		}
+		public IEnumerator<T> GetEnumerator() => _items.GetGenericEnumerator();
 
-		IEnumerator IEnumerable.GetEnumerator() {
-			return _items.GetEnumerator();
-		}
+		IEnumerator IEnumerable.GetEnumerator() => _items.GetEnumerator();
 
 		public static ReadOnlyArray<T> WrapCollection(T[] items) {
 			if (items == null)
 				throw new ArgumentNullException(nameof(items));
 
-			var readOnlyArray = new ReadOnlyArray<T>(items);
-			return readOnlyArray;
+			return new ReadOnlyArray<T>(items);
 		}
 	}
 }
